requirements.txt
Django==4.2.9
django-allauth==0.61.1
django-debug-toolbar==4.2.0
django-environ==0.11.2
psycopg2==2.9.9
django-environ==0.11.2
djangorestframework==3.14.0
drf-yasg==1.21.7
django-cors-headers==4.3.1